move FieldProps to util
