+++ apps/newsletters-ui/src/app/components/SchemaForm/NumberInput.tsx
@@ -1,6 +1,6 @@
 import { css, TextField } from '@mui/material';
 import type { FormEventHandler, FunctionComponent } from 'react';
-import type { FieldProps } from './FieldWrapper';
+import type { FieldProps } from './util';
 import { eventToNumber } from './util';
